requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-dotenv==1.4.2
django-jazzmin==2.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
google-auth==2.28.2
idna==3.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
